linkit: support build using public sdk
This provides support to build lws using the linkit 7697 public SDK
from here https://docs.labs.mediatek.com/resource/mt7687-mt7697/en/downloads

This toolchain has some challenges, its int32_t / uint32_t are long,
so assumptions about format strings for those being %u / %d / %x all
break.  This fixes all the cases for the features enabled by the
default cmake settings.

+++ lib/core-net/wsi.c
@@ -1168,7 +1168,7 @@ lws_wsi_txc_check_skint(struct lws_tx_credit *txc, int32_t tx_cr)
 
 		if (!txc->skint)
 			lwsl_info("%s: %p: skint (%d)\n", __func__, txc,
-				  txc->tx_cr);
+				  (int)txc->tx_cr);
 
 		txc->skint = 1;
 
@@ -1176,7 +1176,8 @@ lws_wsi_txc_check_skint(struct lws_tx_credit *txc, int32_t tx_cr)
 	}
 
 	if (txc->skint)
-		lwsl_info("%s: %p: unskint (%d)\n", __func__, txc, txc->tx_cr);
+		lwsl_info("%s: %p: unskint (%d)\n", __func__, txc,
+			  (int)txc->tx_cr);
 
 	txc->skint = 0;
 
@@ -1188,8 +1189,8 @@ void
 lws_wsi_txc_describe(struct lws_tx_credit *txc, const char *at, uint32_t sid)
 {
 	lwsl_info("%s: %p: %s: sid %d: %speer-to-us: %d, us-to-peer: %d\n",
-		  __func__, txc, at, sid, txc->skint ? "SKINT, " : "",
-		  txc->peer_tx_cr_est, txc->tx_cr);
+		  __func__, txc, at, (int)sid, txc->skint ? "SKINT, " : "",
+		  (int)txc->peer_tx_cr_est, (int)txc->tx_cr);
 }
 #endif
