+++ lib/roles/ws/ops-ws.c
@@ -513,7 +513,7 @@ spill:
 
 		case LWSWSOPC_PING:
 			lwsl_info("received %d byte ping, sending pong\n",
-						 wsi->ws->rx_ubuf_head);
+						 (int)wsi->ws->rx_ubuf_head);
 
 			if (wsi->ws->ping_pending_flag) {
 				/*
